feat: refactor p3
